Fix issue with travis running force pushed commits (#9)
TravisCi provides the $TRAVIS_COMMIT environment variable which I was
using to get composer to install the correct version. However it appears
that under some circumstances travis does not set this variable to the
correct commit hash, which can cause our builds to fail. This avoids
using that variable all together to avoid that problem.