Update JSONLD module to use new service
As per these tickets:
https://www.drupal.org/node/2758897
https://www.drupal.org/node/2854830

The service `rest.link_manager` has been moved into the HAL module.
This means JSONLD needs to depend on the HAL module, and some class
names etc need to be updated, as well at the wiring for our service.
This should let the JSONLD module run on D8.3.